+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/actions/OxlintFixAllOnSaveAction.kt
@@ -1,8 +1,8 @@
 package com.github.oxc.project.oxcintellijplugin.actions
 
 import com.github.oxc.project.oxcintellijplugin.OxcBundle
-import com.github.oxc.project.oxcintellijplugin.services.OxcServerService
-import com.github.oxc.project.oxcintellijplugin.settings.OxcSettings
+import com.github.oxc.project.oxcintellijplugin.services.OxlintServerService
+import com.github.oxc.project.oxcintellijplugin.settings.OxlintSettings
 import com.intellij.ide.actionsOnSave.impl.ActionsOnSaveFileDocumentManagerListener
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
@@ -12,9 +12,9 @@ import com.intellij.openapi.project.Project
 import com.intellij.platform.ide.progress.runWithModalProgressBlocking
 import kotlinx.coroutines.withTimeout
 
-class OxcFixAllOnSaveAction : ActionsOnSaveFileDocumentManagerListener.ActionOnSave() {
+class OxlintFixAllOnSaveAction : ActionsOnSaveFileDocumentManagerListener.ActionOnSave() {
     override fun isEnabledForProject(project: Project): Boolean {
-        return OxcSettings.getInstance(project).fixAllOnSave
+        return OxlintSettings.getInstance(project).fixAllOnSave
     }
 
     override fun processDocuments(project: Project,
@@ -22,23 +22,23 @@ class OxcFixAllOnSaveAction : ActionsOnSaveFileDocumentManagerListener.ActionOnS
         val notificationGroup = NotificationGroupManager.getInstance().getNotificationGroup("Oxc")
 
         runWithModalProgressBlocking(project,
-            OxcBundle.message("oxc.run.fix.all")) {
+            OxcBundle.message("oxlint.run.fix.all")) {
             try {
                 withTimeout(5_000) {
                     documents.filter {
-                        val settings = OxcSettings.getInstance(project)
+                        val settings = OxlintSettings.getInstance(project)
                         val manager = FileDocumentManager.getInstance()
                         val virtualFile = manager.getFile(it) ?: return@filter false
                         return@filter settings.fileSupported(virtualFile)
                     }.forEach {
-                        OxcServerService.getInstance(project).fixAll(it)
+                        OxlintServerService.getInstance(project).fixAll(it)
                     }
                 }
             } catch (e: Exception) {
                 notificationGroup.createNotification(
-                    title = OxcBundle.message("oxc.fix.all.on.save.failure.label"),
+                    title = OxcBundle.message("oxlint.fix.all.on.save.failure.label"),
                     content = OxcBundle.message(
-                        "oxc.fix.all.on.save.failure.description",
+                        "oxlint.fix.all.on.save.failure.description",
                         e.message.toString()
                     ),
                     type = NotificationType.ERROR).notify(project)