+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/listeners/OxlintEditorWatcher.kt
@@ -1,25 +1,25 @@
 package com.github.oxc.project.oxcintellijplugin.listeners
 
-import com.github.oxc.project.oxcintellijplugin.services.OxcServerService
-import com.github.oxc.project.oxcintellijplugin.settings.OxcSettings
+import com.github.oxc.project.oxcintellijplugin.services.OxlintServerService
+import com.github.oxc.project.oxcintellijplugin.settings.OxlintSettings
 import com.intellij.openapi.fileEditor.FileEditorManager
 import com.intellij.openapi.fileEditor.FileEditorManagerListener
 import com.intellij.openapi.vfs.VirtualFile
 
-class EditorWatcher : FileEditorManagerListener {
+class OxlintEditorWatcher : FileEditorManagerListener {
 
     override fun fileClosed(source: FileEditorManager, file: VirtualFile) {
         val project = source.project
         if (source.allEditors.isEmpty()) {
-            OxcServerService.getInstance(project).stopServer()
+            OxlintServerService.getInstance(project).stopServer()
             return
         }
 
         val stillHasSupportedFileOpen = source.allEditors.any {
-            OxcSettings.getInstance(project).fileSupported(it.file)
+            OxlintSettings.getInstance(project).fileSupported(it.file)
         }
         if (!stillHasSupportedFileOpen) {
-            OxcServerService.getInstance(project).stopServer()
+            OxlintServerService.getInstance(project).stopServer()
         }
     }