+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/lsp/OxlintLspServerSupportProvider.kt
@@ -1,9 +1,9 @@
 package com.github.oxc.project.oxcintellijplugin.lsp
 
 import com.github.oxc.project.oxcintellijplugin.OxcIcons
-import com.github.oxc.project.oxcintellijplugin.OxcPackage
-import com.github.oxc.project.oxcintellijplugin.settings.OxcConfigurable
-import com.github.oxc.project.oxcintellijplugin.settings.OxcSettings
+import com.github.oxc.project.oxcintellijplugin.OxlintPackage
+import com.github.oxc.project.oxcintellijplugin.settings.OxlintConfigurable
+import com.github.oxc.project.oxcintellijplugin.settings.OxlintSettings
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectRootManager
@@ -14,33 +14,33 @@ import com.intellij.platform.lsp.api.LspServerSupportProvider
 import com.intellij.platform.lsp.api.lsWidget.LspServerWidgetItem
 import kotlin.io.path.Path
 
-class OxcLspServerSupportProvider : LspServerSupportProvider {
+class OxlintLspServerSupportProvider : LspServerSupportProvider {
     override fun fileOpened(project: Project,
         file: VirtualFile,
         serverStarter: LspServerSupportProvider.LspServerStarter) {
         thisLogger().debug("Handling fileOpened for ${file.path}")
 
-        if (!OxcSettings.getInstance(project).fileSupported(file)) {
+        if (!OxlintSettings.getInstance(project).fileSupported(file)) {
             return
         }
 
-        val oxc = OxcPackage(project)
-        if (!oxc.isEnabled()) {
+        val oxlintPackage = OxlintPackage(project)
+        if (!oxlintPackage.isEnabled()) {
             return
         }
-        val executable = oxc.binaryPath(file) ?: return
-        val nodePackage = oxc.getPackage(file)
+        val executable = oxlintPackage.binaryPath(file) ?: return
+        val nodePackage = oxlintPackage.getPackage(file)
         val root = if (nodePackage != null) {
             VirtualFileManager.getInstance().findFileByNioPath(Path(nodePackage.systemIndependentPath))?.parent?.parent ?: return
         } else {
             ProjectRootManager.getInstance(project).fileIndex.getContentRootForFile(file) ?: return
         }
 
-        serverStarter.ensureServerStarted(OxcLspServerDescriptor(project, root, executable))
+        serverStarter.ensureServerStarted(OxlintLspServerDescriptor(project, root, executable))
     }
 
     override fun createLspServerWidgetItem(lspServer: LspServer,
         currentFile: VirtualFile?): LspServerWidgetItem {
-        return LspServerWidgetItem(lspServer, currentFile, OxcIcons.OxcRound, OxcConfigurable::class.java)
+        return LspServerWidgetItem(lspServer, currentFile, OxcIcons.OxcRound, OxlintConfigurable::class.java)
     }
 }