+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/settings/ConfigurationMode.kt
@@ -1,0 +1,7 @@
+package com.github.oxc.project.oxcintellijplugin.settings
+
+enum class ConfigurationMode {
+    DISABLED,
+    AUTOMATIC,
+    MANUAL
+}