+++ src/main/kotlin/com/github/oxc/project/oxcintellijplugin/settings/OxcOnSaveInfoProvider.kt
@@ -7,10 +7,9 @@ import com.intellij.ide.actionsOnSave.ActionOnSaveInfoProvider
 
 class OxcOnSaveInfoProvider : ActionOnSaveInfoProvider() {
     override fun getActionOnSaveInfos(context: ActionOnSaveContext): List<ActionOnSaveInfo> =
-        listOf(OxcOnSaveFixAllActionInfo(context))
+        listOf(OxlintOnSaveFixAllActionInfo(context))
 
     override fun getSearchableOptions(): Collection<String> {
-        return listOf(OxcBundle.message("oxc.fix.all.on.save.checkbox.on.actions.on.save.page"))
+        return listOf(OxcBundle.message("oxlint.fix.all.on.save.checkbox.on.actions.on.save.page"))
     }
 }
-