+++ src/test/kotlin/com/github/oxc/project/oxcintellijplugin/oxlint/CustomConfigHighlightingTest.kt
@@ -1,6 +1,8 @@
-package com.github.oxc.project.oxcintellijplugin
+package com.github.oxc.project.oxcintellijplugin.oxlint
 
 import com.github.oxc.project.oxcintellijplugin.extensions.configureByFileAndCheckLanguageServerHighlighting
+import com.github.oxc.project.oxcintellijplugin.settings.ConfigurationMode
+import com.github.oxc.project.oxcintellijplugin.settings.OxlintSettings
 import com.intellij.testFramework.TestDataPath
 import com.intellij.testFramework.builders.ModuleFixtureBuilder
 import com.intellij.testFramework.fixtures.CodeInsightFixtureTestCase
@@ -8,24 +10,22 @@ import com.intellij.testFramework.fixtures.ModuleFixture
 import com.intellij.testFramework.fixtures.impl.CodeInsightTestFixtureImpl
 
 @TestDataPath("\$CONTENT_ROOT/testData/highlighting/lint")
-class NestedConfigHighlightingTest :
+class CustomConfigHighlightingTest :
     CodeInsightFixtureTestCase<ModuleFixtureBuilder<ModuleFixture>>() {
 
     override fun setUp() {
         super.setUp()
         (myFixture as CodeInsightTestFixtureImpl).canChangeDocumentDuringHighlighting(true)
         myFixture.testDataPath = "src/test/testData/highlighting/lint"
+
+        val oxlintSettings = OxlintSettings.getInstance(myFixture.project)
+        oxlintSettings.configPath = "${myFixture.tempDirPath}/custom-oxlint.jsonc"
+        oxlintSettings.configurationMode = ConfigurationMode.MANUAL
     }
 
     fun testRootFileHighlighting() {
-        myFixture.copyDirectoryToProject("nested-config", "")
+        myFixture.copyDirectoryToProject("custom-config", "")
 
         myFixture.configureByFileAndCheckLanguageServerHighlighting("index.expected.js")
-    }
-
-    fun testSubdirectoryFileHighlighting() {
-        myFixture.copyDirectoryToProject("nested-config", "")
-
-        myFixture.configureByFileAndCheckLanguageServerHighlighting("subdirectory/index.expected.js")
     }
 }