+++ src/test/kotlin/com/github/oxc/project/oxcintellijplugin/oxlint/NestedConfigHighlightingTest.kt
@@ -1,8 +1,6 @@
-package com.github.oxc.project.oxcintellijplugin
+package com.github.oxc.project.oxcintellijplugin.oxlint
 
 import com.github.oxc.project.oxcintellijplugin.extensions.configureByFileAndCheckLanguageServerHighlighting
-import com.github.oxc.project.oxcintellijplugin.settings.ConfigurationMode
-import com.github.oxc.project.oxcintellijplugin.settings.OxcSettings
 import com.intellij.testFramework.TestDataPath
 import com.intellij.testFramework.builders.ModuleFixtureBuilder
 import com.intellij.testFramework.fixtures.CodeInsightFixtureTestCase
@@ -10,22 +8,24 @@ import com.intellij.testFramework.fixtures.ModuleFixture
 import com.intellij.testFramework.fixtures.impl.CodeInsightTestFixtureImpl
 
 @TestDataPath("\$CONTENT_ROOT/testData/highlighting/lint")
-class CustomConfigHighlightingTest :
+class NestedConfigHighlightingTest :
     CodeInsightFixtureTestCase<ModuleFixtureBuilder<ModuleFixture>>() {
 
     override fun setUp() {
         super.setUp()
         (myFixture as CodeInsightTestFixtureImpl).canChangeDocumentDuringHighlighting(true)
         myFixture.testDataPath = "src/test/testData/highlighting/lint"
-
-        val oxcSettings = OxcSettings.getInstance(myFixture.project)
-        oxcSettings.configPath = "${myFixture.tempDirPath}/custom-oxlint.jsonc"
-        oxcSettings.configurationMode = ConfigurationMode.MANUAL
     }
 
     fun testRootFileHighlighting() {
-        myFixture.copyDirectoryToProject("custom-config", "")
+        myFixture.copyDirectoryToProject("nested-config", "")
 
         myFixture.configureByFileAndCheckLanguageServerHighlighting("index.expected.js")
+    }
+
+    fun testSubdirectoryFileHighlighting() {
+        myFixture.copyDirectoryToProject("nested-config", "")
+
+        myFixture.configureByFileAndCheckLanguageServerHighlighting("subdirectory/index.expected.js")
     }
 }