+++ src/test/kotlin/com/github/oxc/project/oxcintellijplugin/oxlint/NoConfigHighlightingTest.kt
@@ -1,4 +1,4 @@
-package com.github.oxc.project.oxcintellijplugin
+package com.github.oxc.project.oxcintellijplugin.oxlint
 
 import com.github.oxc.project.oxcintellijplugin.extensions.configureByFileAndCheckLanguageServerHighlighting
 import com.intellij.testFramework.TestDataPath